requirements.txt
pysqlite3-binary
python-dotenv==1.0.1
pypdf==4.2.0
chromadb==0.4.24
tiktoken==0.6.0
langchain==0.1.16
openai==1.17.0
streamlit==1.33.0
streamlit-pills==0.3.0
streamlit-option-menu==0.3.12
langchain-community==0.0.32
langchain-openai==0.1.3